requirements.txt
customtkinter>=5.2.0
pyautogui>=0.9.54
pytesseract>=0.3.10
Pillow>=10.0.0
openpyxl>=3.1.2
pymysql>=1.1.0
